Minor changes to volume.md
This patch adds additional content
to the description of the hostPath
volume type based on things that I
did not find to be clear from the
documentation.

1. included an example pod definition
in the hostPath description to demonstrate
that `path` is the name of the key to use to
specify the local directory location.

2. documented that any directories created
based on hostPath definitions will only be writable
by root and offers some ways of working around this.