Use subPath to keep data dir out of volume root.
The data dir should not be at the root of the volume, because many
volume types create files or directories there automatically.
The database initialization script requires the data dir to be empty,
and we shouldn't try to guess which files may exist.

In particular, by retaining lost+found, it also becomes possible
to recover files that end up there.