hello-minikube: Simplify kubectl installation instructions for macOS
People are hitting issues copy-pasting the complicated way to install
`kubectl` and we're suggesting them to `sudo mv` binaries to /usr/local/bin.
That is not very ideal. Instead suggesting users to use Homebrew. I think it's
ok for this tutorial, as the tutorial already uses Homebrew to install
other stuff, plus it's a much better way than `sudo mv`ing.

Closes #2478.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>