Change example PersistentVolumeClaim
If a default storage class exists, it is going to create a new persistent volume to bind to that claim instead of using our example one, which will make the tutorial fail. In particular, this happens on minikube.

Specifying an empty storageClassName will prevent the storage class to be used, and force our volume to be selected.