Be more specific about cpu-shares passed to Docker (#4016)
Docker documentation (https://docs.docker.com/engine/reference/run/) reads:
  
"To modify the proportion from the default of 1024, use the -c or --cpu-shares flag to set the weighting to 2 or higher. If 0 is set, the system will ignore the value and use the default of 1024."

So the Kubernetes documentation gives the impression that if requests.cpu is zero, Docker would end up applying a value of 1024, which isn't the case.

This change makes the explanation more precise, hopefully without making it much more difficult to read.