Script to import tutorials from examples repo
This Python script imports the tutorials that are going to be maintained
at the https://github.com/kubernetes/examples repository. Currently only
pulls the tutorials from master ref of examples.

This script:

- Deletes content between `<!-- EXCLUDE_FROM_DOCS BEGIN -->` and
  `<!-- EXCLUDE_FROM_DOCS END -->` (including the markers). This is to drop
  website redirection notices appear in the examples repo.
- Adds Jekyll title to the imported docs. The title can be customized so we do
  not need to parse h1 elements from the markdown.
- Insert "DO NOT UPDATE" section (commented out HTML) to the markdown files
  that are imported.

Once merged, I will go ahead and update the tutorials in kubernetes/examples,
then I will run this script and check in its results in a separate PR.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>