Update on install-kubectl.md (#6954)
* Update on install-kubectl.md

Added more info on the reason people may get "the connection to the server was refused error". It will help newbies understand the kubectl is a client that connects to a cluster. Hence, a Kubernetes cluster such as a minikube tool will be needed to verify that kubectl is functioning correctly.

* updating wording to remove redundancy