Avoid misleading camel words (#8075)
The term "`MutatingAdmissionWebhooks`" has no special meaning in the
docs. It is misleading because readers may think of it as the name of a
resource or field, or the name of the corresponding admission
controller. However, the name of the admission controller is in
singular. This PR proposes a change to avoid such misunderstandings.