Delete docs for ThirdPartyResources (#8544)
ThirdPartyResources (TPRs) were removed in v1.8. It is misleading
to still show the docs for TPRs for > v1.8. If one needs to read
docs for TPRs, they can switch to the v1.8 version of the docs
website.

However, the docs of migration of TPRs to CRDs are still not
removed because one can also migrate from v1.7 (when CRDs were
introduced) to v1.7+.