Clarify that beta annotations can not be removed (#8666)
* Clarify that beta annotations can not be removed

Removing annotations caused a major outage with nodes that lagged the
master.  We used annotations as would-be fields, though we don't any
more.  Now we just add fields.  That makes them part of the API version.
These annotations are effectively part of the API version, too.

It stinks because I love to delete code, but I think it's just part of
the debt-accrual process and yet another reason we need to be judicious
in adding fields.

* Change 6 month deprecations to 9 months

This covers maximum version skew.  For example a master might be at
1.10, while nodes are at 1.8.  We must not break compat within that
window.