Updated pick-right-solution.md with Docker Enterprise (#12220)
* Updated pick-right-solution.md with Docker Enterprise

Added Docker commercial and community products to the list of solutions for running Kubernetes including Docker Desktop and Docker Enterprise.

* Fix Docker Desktop entry in pick-right-solution.md 

The words "Docker Desktop" was duplicated in the description under locally-hosted solutions.

* Updated pick-right-solution.md with Docker solutions

Added line wraps at 80 characters and changed containerizing to "deploying in containers" to make it more new-user friendly.

* Updated pick-right-solution.md

modified wrapping of first line to not break hyphenated word.