Blog post links (hugo issue) (#18307)
* Blog post links (hugo issue)

This commit aims to fix multiple occurrences of issue introduced before
switching blog engine to hugo, after which no blog links with "." was
working.

It's include a huge ammount of links with kubernetes version in it.

This commit also include one or two additional link fixes found while
making changes in attempt to solve current situation.

* Additional fix for Chinese website blog post.